# api/swagger.yaml
basePath: /
definitions:
...
  model.Date:
    properties:
      day:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
...
  model.Item:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.Recommendation:
    properties:
      recommended:
        $ref: '#/definitions/model.Source'
      source:
        $ref: '#/definitions/model.Source'
      users:
        items:
          $ref: '#/definitions/model.RecommendationUser'
        type: array
    type: object
  model.RecommendationUser:
    properties:
      content:
        type: string
      username:
        type: string
    type: object
...
  model.Review:
    properties:
      chapter:
        description: for manga
        type: string
      date:
        type: string
      episode:
        description: for anime
        type: string
      helpful:
        type: integer
      id:
        type: integer
      image:
        type: string
      review:
        type: string
      score:
        $ref: '#/definitions/model.ReviewScore'
      source:
        $ref: '#/definitions/model.Source'
      username:
        type: string
    type: object
  model.ReviewScore:
    properties:
      art:
        description: for both
        type: integer
      character:
        description: for both
        type: integer
      enjoyment:
        description: for both
        type: integer
      overall:
        description: for both
        type: integer
      sound:
        description: for anime only
        type: integer
      story:
        description: for both
        type: integer
    type: object
...
  model.Source:
    properties:
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  model.User:
    properties:
      about:
        type: string
      birthday:
        $ref: '#/definitions/model.Date'
      blogPost:
        type: integer
      club:
        type: integer
      forumPost:
        type: integer
      friend:
        type: integer
      gender:
        type: string
      image:
        type: string
      joinedDate:
        type: string
      lastOnline:
        type: string
      location:
        type: string
      recommendation:
        type: integer
      review:
        type: integer
      sns:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  model.UserAnime:
    properties:
      airingEnd:
        $ref: '#/definitions/model.Date'
      airingStart:
        $ref: '#/definitions/model.Date'
      airingStatus:
        type: integer
      days:
        type: integer
      episode:
        type: integer
      id:
        type: integer
      image:
        type: string
      isRewatching:
        type: boolean
      priority:
        type: string
      progress:
        type: integer
      rating:
        type: string
      score:
        type: integer
      status:
        type: integer
      storage:
        type: string
      tag:
        type: string
      title:
        type: string
      type:
        type: string
      watchEnd:
        type: string
      watchStart:
        type: string
    type: object
  model.UserAnimeStats:
    properties:
      completed:
        type: integer
      current:
        type: integer
      days:
        type: number
      dropped:
        type: integer
      episode:
        type: integer
      meanScore:
        type: number
      onHold:
        type: integer
      planned:
        type: integer
      rewatched:
        type: integer
      total:
        type: integer
    type: object
  model.UserFavorite:
    properties:
      anime:
        items:
          $ref: '#/definitions/model.UserFavoriteItem'
        type: array
      character:
        items:
          $ref: '#/definitions/model.UserFavoriteItem'
        type: array
      manga:
        items:
          $ref: '#/definitions/model.UserFavoriteItem'
        type: array
      people:
        items:
          $ref: '#/definitions/model.UserFavoriteItem'
        type: array
    type: object
  model.UserFavoriteItem:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
    type: object
  model.UserFriend:
    properties:
      friendSince:
        type: string
      image:
        type: string
      lastOnline:
        type: string
      username:
        type: string
    type: object
  model.UserHistory:
    properties:
      date:
        type: string
      id:
        type: integer
      progress:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  model.UserManga:
    properties:
      chapter:
        type: integer
      chapterProgress:
        type: integer
      days:
        type: integer
      id:
        type: integer
      image:
        type: string
      isRereading:
        type: boolean
      priority:
        type: string
      publishingEnd:
        $ref: '#/definitions/model.Date'
      publishingStart:
        $ref: '#/definitions/model.Date'
      publishingStatus:
        type: integer
      readEnd:
        type: string
      readStart:
        type: string
      score:
        type: integer
      status:
        type: integer
      tag:
        type: string
      title:
        type: string
      type:
        type: string
      volume:
        type: integer
      volumeProgress:
        type: integer
    type: object
  model.UserMangaStats:
    properties:
      chapter:
        type: integer
      completed:
        type: integer
      current:
        type: integer
      days:
        type: number
      dropped:
        type: integer
      meanScore:
        type: number
      onHold:
        type: integer
      planned:
        type: integer
      reread:
        type: integer
      total:
        type: integer
      volume:
        type: integer
    type: object
...
  model.UserStats:
    properties:
      anime:
        $ref: '#/definitions/model.UserAnimeStats'
      manga:
        $ref: '#/definitions/model.UserMangaStats'
    type: object
...
  utils.Response:
    properties:
      data:
        type: object
      message:
        type: string
      meta:
        type: object
      status:
        type: integer
    type: object
info:
  contact:
    email: axel.rl.404@gmail.com
    name: Axel
    url: https://github.com/rl404
  description: Scraping/parsing MyAnimeList website to a useful and easy-to-use data.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Malscraper API
paths:
...
  /user/{user}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
      summary: Get user details
      tags:
      - user
  /user/{user}/anime:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Progress status (1=watching, 2=completed, 3=on-hold, 4=dropped, 6=planned, 7=all)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 6
        - 7
        in: query
        name: status
        type: integer
      - description: Order (1=title, 2=finish date, 3=start date, 4=score, 6=type, 8=rating, 11=priority, 12=progress, 13=storage, 14=airing start, 15=airing end)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 6
        - 8
        - 11
        - 12
        - 13
        - 14
        - 15
        in: query
        name: order
        type: integer
      - description: Tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserAnime'
                  type: array
              type: object
      summary: Get user anime list
      tags:
      - user
  /user/{user}/clubs:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Item'
                  type: array
              type: object
      summary: Get user club list
      tags:
      - user
  /user/{user}/favorites:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserFavorite'
              type: object
      summary: Get user favorite list
      tags:
      - user
  /user/{user}/friends:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserFriend'
                  type: array
              type: object
      summary: Get user friend list
      tags:
      - user
  /user/{user}/history:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserHistory'
                  type: array
              type: object
      summary: Get user anime & manga history
      tags:
      - user
  /user/{user}/history/anime:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserHistory'
                  type: array
              type: object
      summary: Get user anime history
      tags:
      - user
  /user/{user}/history/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserHistory'
                  type: array
              type: object
      summary: Get user manga history
      tags:
      - user
  /user/{user}/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Progress status (1=watching, 2=completed, 3=on-hold, 4=dropped, 6=planned, 7=all)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 6
        - 7
        in: query
        name: status
        type: integer
      - description: Order (1=title, 2=finish date, 3=start date, 4=score, 7=priority, 8=chapter, 9=volume, 10=type, 11=publishing start, 12=publishing end)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 7
        - 8
        - 9
        - 10
        - 11
        - 12
        in: query
        name: order
        type: integer
      - description: Tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserManga'
                  type: array
              type: object
      summary: Get user manga list
      tags:
      - user
  /user/{user}/recommendations:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Recommendation'
                  type: array
              type: object
      summary: Get user recommendation list
      tags:
      - user
  /user/{user}/reviews:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Review'
                  type: array
              type: object
      summary: Get user review list
      tags:
      - user
  /user/{user}/stats:
    get:
      consumes:
      - application/json
      parameters:
      - description: User name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.UserStats'
              type: object
      summary: Get user stats
      tags:
      - user
schemes:
- http
- https
swagger: "2.0"